# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WhatsApp Reverse Auth API",
    "version": "1.0.0",
    "description": "Interactive documentation for the WhatsApp automation system."
  },
  "paths": {
...
    "/api/auth/challenge": {
      "get": {
        "summary": "Check challenge status",
        "tags": ["Auth"],
        "parameters": [
          { "name": "token", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Challenge status returned" },
          "404": { "description": "Challenge not found or expired" }
        }
      },
      "post": {
        "summary": "Generate an authentication challenge",
        "tags": ["Auth"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "metadata": { "type": "object", "description": "Optional data to attach" },
                  "ttl": { "type": "integer", "default": 600, "description": "Time-to-live in seconds" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": { "type": "string" },
                    "whatsapp_url": { "type": "string", "nullable": true },
                    "bot_number": { "type": "string", "nullable": true },
                    "expiresAt": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}